+++ app/page.js
@@ -5,53 +5,19 @@ import oscardev from '../public/ovg-avatar.png';
 import languages from '../public/languages1.jpg';
 import framework from '../public/framework.jpg';
 import skills from '../public/skills.jpg';
-import project1 from '../public/project1.png';
-import project2 from '../public/project2.png';
-import project3 from '../public/project3.png';
-import project4 from '../public/project4.png';
-import web6 from '../public/EasyClinics-Appointment.png';
 import '@tailwindcss/forms';
-import formImage from '../public/form.png';
-import { useFormik } from 'formik';
-import * as Yup from 'yup';
-import { useRouter } from 'next/navigation';
-import { motion as m } from 'framer-motion';
-import Navbar from './components/navbar';
+import Navbar from './components/Navbar';
 import { useDarkMode } from './DarkModeContext';
 import Connect from './components/Connect';
-import {AiFillGithub, AiFillHtml5} from 'react-icons/ai';
+import {AiFillHtml5} from 'react-icons/ai';
 import {BiLogoCss3, BiLogoJavascript, BiLogoPostgresql, BiLogoTailwindCss, BiLogoReact, BiLogoRedux} from 'react-icons/bi';
 import {DiRuby, DiBootstrap} from 'react-icons/di';
 import {SiRubyonrails} from 'react-icons/si';
-import {GiNothingToSay} from 'react-icons/gi';
-import {LuArrowDownRightFromCircle} from 'react-icons/lu'
+import Cards from './components/Cards';
+import Form from './components/Form';
 
 export default function Home() {
   const { darkMode } = useDarkMode();
-
-  // Router
-  const router = useRouter();
-
-  // Formic Logics
-  const formik = useFormik({
-    initialValues: {
-      name: '',
-      email: '',
-      text: '',
-    },
-
-    // validate form
-    validationSchema: Yup.object({
-      name: Yup.string().max(20, 'Name must be 20 characters or less').required('Name is Required'),
-      email: Yup.string().email('Invalid email address').required('Email is Required'),
-      text: Yup.string().max(300, 'Message must be 300 characters or less').required('Message is Required'),
-    }),
-
-    // Submit form
-    onSubmit: (values) => {
-      router.push('/success');
-    },
-  });
 
   return (
     <div className={darkMode ? "dark" : ""}>
@@ -92,289 +58,7 @@ export default function Home() {
               I offer from a wide range of services, including programming and teaching.
             </p>
           </div>
-          <div className='flex flex-col gap-10 py-10 md:flex-row md:flex-wrap min-h-screen items-center justify-center rounded-xl bg-slate-100 dark:bg-gray-600'>
-            <div className='group h-96 w-96 [perspective:1000px]'>
-              <div className='relative h-full w-full rounded-xl shadow-xl transition-all duration-500 [transform-style:preserve-3d] group-hover:[transform:rotateY(360deg)]'>
-                <div className='absolute inset-0'>
-                  <Image src={project1} alt='project-image' className='w-full h-full rounded-xl object-cover shadow-xl shadow-black/40' />
-                </div>
-                <div className='absolute inset-0 h-full w-full rounded-xl bg-black/60 px-12 text-center text-slate-200 [transform:rotateY(360deg)] transition-opacity opacity-0 group-hover:opacity-100'>
-                  <div className='flex min-h-full flex-col items-center justify-center'>
-                    <h1 className='text-3xl font-bold text-teal-200'>
-                      Global Submit Conference
-                    </h1>
-                    <p className='text-lg'>
-                      Conference Web Page
-                    </p>
-                    <p className='text-base'>
-                      This is an HTML/CSS & JavaScript capstone project for the Full-Stack Web Development Program at Microverse. This capstone project is about a conference page.
-                    </p>
-                    <div className='flex justify-between w-full'>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://oovillagran.github.io/Global-Summit-Conference-/' target='_blank' rel='noopener noreferrer'>
-                          See Live
-                        </a>
-                        <LuArrowDownRightFromCircle className='inline-block ml-2 text-xl' />
-                      </button>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://github.com/oovillagran/Global-Summit-Conference-' target='_blank' rel='noopener noreferrer'>
-                          See Source
-                        </a>
-                        <AiFillGithub className='inline-block ml-2 text-xl' />
-                      </button>
-                    </div>
-                    <div className='flex items-center flex-wrap py-5 justify-center'>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        HTML
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        CSS
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        JavaScript
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className='group h-96 w-96 [perspective:1000px]'>
-              <div className='relative h-full w-full rounded-xl shadow-xl transition-all duration-500 [transform-style:preserve-3d] group-hover:[transform:rotateY(360deg)]'>
-                <div className='absolute inset-0'>
-                  <Image src={project2} alt='project-image' className='w-full h-full rounded-xl object-cover shadow-xl shadow-black/40' />
-                </div>
-                <div className='absolute inset-0 h-full w-full rounded-xl bg-black/60 px-12 text-center text-slate-200 [transform:rotateY(360deg)] transition-opacity opacity-0 group-hover:opacity-100'>
-                  <div className='flex min-h-full flex-col items-center justify-center'>
-                    <h1 className='text-3xl font-bold text-teal-200'>
-                      Central Cinema Trends
-                    </h1>
-                    <p className='text-lg'>
-                      FrontEnd App
-                    </p>
-                    <p className='text-base'>
-                      Central Cinema Trends is a webapp that shows you the most popular movies at the moment.
-                    </p>
-                    <div className='flex justify-between w-full'>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://centralcinematrends.onrender.com/' target='_blank' rel='noopener noreferrer'>
-                          See Live
-                        </a>
-                        <LuArrowDownRightFromCircle className='inline-block ml-2 text-xl' />
-                      </button>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://github.com/oovillagran/Central-Cinema-Trends' target='_blank' rel='noopener noreferrer'>
-                          See Source
-                        </a>
-                        <AiFillGithub className='inline-block ml-2 text-xl' />
-                      </button>
-                    </div>
-                    <div className='flex items-center flex-wrap py-5 justify-center'>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        HTML
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        CSS
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        JavaScript
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        API-REST
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className='group h-96 w-96 [perspective:1000px]'>
-              <div className='relative h-full w-full rounded-xl shadow-xl transition-all duration-500 [transform-style:preserve-3d] group-hover:[transform:rotateY(360deg)]'>
-                <div className='absolute inset-0'>
-                  <Image src={project3} alt='project-image' className='w-full h-full rounded-xl object-cover shadow-xl shadow-black/40' />
-                </div>
-                <div className='absolute inset-0 h-full w-full rounded-xl bg-black/60 px-12 text-center text-slate-200 [transform:rotateY(360deg)] transition-opacity opacity-0 group-hover:opacity-100'>
-                  <div className='flex min-h-full flex-col items-center justify-center'>
-                    <h1 className='text-3xl font-bold text-teal-200'>
-                      Movie's Time App
-                    </h1>
-                    <p className='text-lg'>
-                      FrontEnd
-                    </p>
-                    <p className='text-base'>
-                      Designed for movie lovers. Powered by React and Redux, this application categorizes the top trending movies into four sections.
-                    </p>
-                    <div className='flex justify-between w-full'>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://movies-time.onrender.com/' target='_blank' rel='noopener noreferrer'>
-                          See Live
-                        </a>
-                        <LuArrowDownRightFromCircle className='inline-block ml-2 text-xl' />
-                      </button>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://github.com/oovillagran/Movie-s-time-app' target='_blank' rel='noopener noreferrer'>
-                          See Source
-                        </a>
-                        <AiFillGithub className='inline-block ml-2 text-xl' />
-                      </button>
-                    </div>
-                    <div className='flex items-center flex-wrap py-5 justify-center'>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        React
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Redux
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        API-REST
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        JavaScript
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className='group h-96 w-96 [perspective:1000px]'>
-              <div className='relative h-full w-full rounded-xl shadow-xl transition-all duration-500 [transform-style:preserve-3d] group-hover:[transform:rotateY(360deg)]'>
-                <div className='absolute inset-0'>
-                  <Image src={project4} alt='project-image' className='w-full h-full rounded-xl object-cover shadow-xl shadow-black/40' />
-                </div>
-                <div className='absolute inset-0 h-full w-full rounded-xl bg-black/60 px-12 text-center text-slate-200 [transform:rotateY(360deg)] transition-opacity opacity-0 group-hover:opacity-100'>
-                  <div className='flex min-h-full flex-col items-center justify-center'>
-                    <h1 className='text-3xl font-bold text-teal-200'>
-                      MyXpenses
-                    </h1>
-                    <p className='text-lg'>
-                      FrontEnd and BackEnd App
-                    </p>
-                    <p className='text-base'>
-                      This is a mobile web application that allows you to manage your budget, organize your list of daily expenses and take control over your incomes/expenses.
-                    </p>
-                    <div className='flex justify-between w-full'>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://rails-hc3r.onrender.com' target='_blank' rel='noopener noreferrer'>
-                          See Live
-                        </a>
-                        <LuArrowDownRightFromCircle className='inline-block ml-2 text-xl' />
-                      </button>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://github.com/oovillagran/myxpenses' target='_blank' rel='noopener noreferrer'>
-                          See Source
-                        </a>
-                        <AiFillGithub className='inline-block ml-2 text-xl' />
-                      </button>
-                    </div>
-                    <div className='flex items-center flex-wrap py-5 justify-center'>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Ruby
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Rails
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        PostgreSQL
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Full-Stack
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className='group h-96 w-96 [perspective:1000px]'>
-              <div className='relative h-full w-full rounded-xl shadow-xl transition-all duration-500 [transform-style:preserve-3d] group-hover:[transform:rotateY(360deg)]'>
-                <div className='absolute inset-0'>
-                  <Image src={web6} className='w-full h-full rounded-xl object-cover shadow-xl shadow-black/40' />
-                </div>
-                <div className='absolute inset-0 h-full w-full rounded-xl bg-black/60 px-12 text-center text-slate-200 [transform:rotateY(360deg)] transition-opacity opacity-0 group-hover:opacity-100'>
-                  <div className='flex min-h-full flex-col items-center justify-center'>
-                    <h1 className='text-3xl font-bold text-teal-200'>
-                      Easy Clinics
-                    </h1>
-                    <p className='text-lg'>
-                      FrontEnd and BackEnd App
-                    </p>
-                    <p className='text-base'>
-                      This awesome app allows you to make an appointment with your favorite doctor in your preferred city.
-                    </p>
-                    <div className='flex justify-between w-full'>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        See Live
-                        <LuArrowDownRightFromCircle className='inline-block ml-2 text-xl' />
-                      </button>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        See Source
-                        <AiFillGithub className='inline-block ml-2 text-xl' />
-                      </button>
-                    </div>
-                    <div className='flex items-center flex-wrap py-5 justify-center'>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        React
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Redux
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Rails
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        PostgreSQL
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-            <div className='group h-96 w-96 [perspective:1000px]'>
-              <div className='relative h-full w-full rounded-xl shadow-xl transition-all duration-500 [transform-style:preserve-3d] group-hover:[transform:rotateY(360deg)]'>
-                <div className='absolute inset-0'>
-                  <Image src={web6} className='w-full h-full rounded-xl object-cover shadow-xl shadow-black/40' />
-                </div>
-                <div className='absolute inset-0 h-full w-full rounded-xl bg-black/60 px-12 text-center text-slate-200 [transform:rotateY(360deg)] transition-opacity opacity-0 group-hover:opacity-100'>
-                  <div className='flex min-h-full flex-col items-center justify-center'>
-                    <h1 className='text-3xl font-bold text-teal-200'>
-                      Easy Clinics
-                    </h1>
-                    <p className='text-lg'>
-                      FrontEnd and BackEnd App
-                    </p>
-                    <p className='text-base'>
-                      This awesome app allows you to make an appointment with your favorite doctor in your preferred city.
-                    </p>
-                    <div className='flex justify-between w-full'>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://easyclinic-client.vercel.app/' target='_blank' rel='noopener noreferrer'>
-                          See Live
-                        </a>
-                        <LuArrowDownRightFromCircle className='inline-block ml-2 text-xl' />
-                      </button>
-                      <button className='mt-2 rounded-md bg-gradient-to-l from-cyan-500 to-teal-600 py-1 px-2 text-sm hover:text-gray-800 transition-transform transform hover:scale-110'>
-                        <a href='https://github.com/oovillagran/final-capstone-back-end' target='_blank' rel='noopener noreferrer'>
-                          See Source
-                        </a>
-                        <AiFillGithub className='inline-block ml-2 text-xl' />
-                      </button>
-                    </div>
-                    <div className='flex items-center flex-wrap py-5 justify-center'>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        React
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Redux
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        Rails
-                      </p>
-                      <p className='text-sm bg-teal-500 px-2 py-1 mr-5 mb-2 rounded-full text-white'>
-                        PostgreSQL
-                      </p>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
+          <Cards />
         </section>
       {/* About me */}
       <section id='about' className='py-1'>
@@ -530,101 +214,7 @@ export default function Home() {
       </main>
 
       {/* Contact section */}
-      <m.div
-        initial={{opacity: 0}}
-        animate={{opacity: 1}}
-        exit={{opacity: 0}}
-        className='absolute w-full'
-      >
-        <section id='contact' className='h-screen flex items-center justify-center bg-gradient-to-r from-cyan-500 to-teal-500'>
-          <form
-            onSubmit={formik.handleSubmit}
-            className='bg-white flex flex-col md:flex-row rounded-lg font-latoRegular shadow-lg w-11/12'
-          >
-            <div className='flex-1 text-gray-700 p-5'>
-              <div className='flex items-center'>
-                <h1 className='text-2xl md:text-3xl font-bold px-2 lg:text-4xl pb-2 font-latoBold text-teal-500'>
-                  Let's connect 
-                </h1>
-                <GiNothingToSay className='text-4xl text-teal-500'/>
-              </div>
-              <p className='text-md text-gray-500 dark:text-gray-200'>
-                If you have an application you are interested in developing, 
-                a feature you need built, or a project that needs coding. I'd 
-                love to help with it. Feel free to reach out to me. 
-                I'll do my best to get back to you!
-              </p>
-              <div className='mt-5'>
-                {/* Name input field */}
-                <div className='pb-4'>
-                  <label
-                    className={`block font-latoBold text-sm pb-2 ${formik.touched.name && formik.errors.name ? 'text-red-500' : ''} dark:text-gray-200`}
-                    htmlFor='name'
-                  >
-                    {formik.touched.name && formik.errors.name ? formik.errors.name : "Name:"}
-                  </label>
-                  <input
-                    className='border-2 border-gray-500 p-2 text-sm rounded-md w-3/4 lg:w-1/2 focus:border-teal-500 focus:ring-teal-500'
-                    type='text'
-                    name='name'
-                    placeholder='Enter your name'
-                    value={formik.values.name}
-                    onChange={formik.handleChange}
-                    onBlur={formik.handleBlur}
-                  />
-                </div>
-                {/* Email input field */}
-                <div className='pb-4'>
-                  <label
-                    className={`block font-latoBold text-sm pb-2 ${formik.touched.email && formik.errors.email ? 'text-red-500' : ''} dark:text-gray-200`}
-                    htmlFor='email'
-                  >
-                    {formik.touched.email && formik.errors.email ? formik.errors.email : "Email:"}
-                  </label>
-                  <input
-                    className='border-2 border-gray-500 p-2 text-sm rounded-md w-3/4 lg:w-1/2 focus:border-teal-500 focus:ring-teal-500'
-                    type='email'
-                    name='email' 
-                    placeholder='Enter your email'
-                    value={formik.values.email}
-                    onChange={formik.handleChange}
-                    onBlur={formik.handleBlur}
-                  />
-                </div>
-                {/* Text area field */}
-                <div className='pb-4'>
-                  <label
-                    className={`block font-latoBold text-sm pb-2 ${formik.touched.text && formik.errors.text ? 'text-red-500' : ''} dark:text-gray-200`}
-                    htmlFor='text'
-                  >
-                    {formik.touched.text && formik.errors.text ? formik.errors.text : "Write your message:"}
-                  </label>
-                  <textarea
-                    className='border-2 border-gray-500 p-2 text-sm rounded-md w-3/4 lg:w-1/2 focus:border-teal-500 focus:ring-teal-500 h-40'
-                    type='textarea'
-                    name='text' 
-                    placeholder='Write your message here...'
-                    style={{ resize: 'none' }}
-                    value={formik.values.text}
-                    onChange={formik.handleChange}
-                    onBlur={formik.handleBlur}
-                  />
-                </div>
-                {/* Submit button */}
-                <button
-                  type='submit'
-                  className='bg-teal-500 font-latoBold text-white px-4 py-2 rounded-md'
-                >
-                  Get in Touch
-                </button>
-              </div>
-            </div>
-            <div className='flex-1 relative'>
-              <Image src={formImage} alt="image" fill className='object-cover rounded-lg' />
-            </div>
-          </form>
-        </section>
-      </m.div>
+      <Form />
     </div>
   );
 }
